agent: Add config to Context
This will be required to call the message relayer with the correct data.
